2.0.0.7.66

+++ samples/generator2/app/src/main/java/com/example/generator2/mp3/channel.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.channels.Channel
 
 val chDataStreamOutAudioProcessor =
-    Channel<ShortArray>(capacity = 16, BufferOverflow.DROP_LATEST)
+    Channel<FloatArray>(capacity = 16, BufferOverflow.DROP_LATEST)
 
 val channelDataStreamOutGenerator =
     Channel<ShortArray>(capacity = 16, BufferOverflow.DROP_OLDEST)
